+++ src/mako/database/connections/MariaDB.php
@@ -1,0 +1,16 @@
+<?php
+
+/**
+ * @copyright Frederic G. Østby
+ * @license   http://www.makoframework.com/license
+ */
+
+namespace mako\database\connections;
+
+/**
+ * MariaDB database connection.
+ */
+class MariaDB extends MySQL
+{
+
+}