+++ src/mako/database/query/compilers/Firebird.php
@@ -79,6 +79,30 @@ class Firebird extends Compiler
 	/**
 	 * {@inheritDoc}
 	 */
+	public function insertAndReturn(array $values, array $return): array
+	{
+		['sql' => $sql, 'params' => $params] = $this->insert($values);
+
+		$sql .= ' RETURNING ' . $this->columnNames($return);
+
+		return ['sql' => $sql, 'params' => $params];
+	}
+
+	/**
+	 * {@inheritDoc}
+	 */
+	public function insertMultipleAndReturn(array $return, array ...$values): array
+	{
+		['sql' => $sql, 'params' => $params] = $this->insertMultiple(...$values);
+
+		$sql .= ' RETURNING ' . $this->columnNames($return);
+
+		return ['sql' => $sql, 'params' => $params];
+	}
+
+	/**
+	 * {@inheritDoc}
+	 */
 	public function updateAndReturn(array $values, array $return): array
 	{
 		$query = $this->update($values);