+++ src/mako/database/query/compilers/MariaDB.php
@@ -1,0 +1,38 @@
+<?php
+
+/**
+ * @copyright Frederic G. Østby
+ * @license   http://www.makoframework.com/license
+ */
+
+namespace mako\database\query\compilers;
+
+/**
+ * Compiles MariaDB queries.
+ */
+class MariaDB extends MySQL
+{
+	/**
+	 * {@inheritDoc}
+	 */
+	public function insertAndReturn(array $values, array $return): array
+	{
+		['sql' => $sql, 'params' => $params] = $this->insert($values);
+
+		$sql .= ' RETURNING ' . $this->columnNames($return);
+
+		return ['sql' => $sql, 'params' => $params];
+	}
+
+	/**
+	 * {@inheritDoc}
+	 */
+	public function insertMultipleAndReturn(array $return, array ...$values): array
+	{
+		['sql' => $sql, 'params' => $params] = $this->insertMultiple(...$values);
+
+		$sql .= ' RETURNING ' . $this->columnNames($return);
+
+		return ['sql' => $sql, 'params' => $params];
+	}
+}